Feature/chat

+++ src/main/java/com/mutsa/springboot_auction/domain/bid/entity/Bid.java
@@ -15,7 +15,7 @@ public class Bid extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer bidId;
+    private Long bidId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "userId", nullable = false)
